(RES): Revisited naming hierarchy: introduced `RustAnonymousId` to designate        Rust's top-level anonymous module-scope
