+++ src/main/kotlin/org/rust/lang/core/names/RustQualifiedName.kt
@@ -1,40 +1,30 @@
 package org.rust.lang.core.names
 
+import org.rust.lang.core.names.parts.RustIdNamePart
+import org.rust.lang.core.names.parts.RustNamePart
+
 /**
  * Abstract qualified-name representation serving purposes of
  * unifying PSI interface with PSI-independent IR
  *
+ * Serves primarily as an URI for items inside the Rust's crates
+ *
  * @name        Non-qualified name-part
  * @qualifier   Qualified name-part
  */
-data class RustQualifiedName(val part: RustNamePart, val qualifier: RustQualifiedName? = null) {
+open class RustQualifiedName(val part: RustNamePart, val qualifier: RustQualifiedName? = null) {
+
+    override fun toString(): String =
+        "${qualifier?.toString()}::${part.toString()}"
+
     companion object {
+
         fun parse(s: String): RustQualifiedName? {
-            val q: RustQualifiedName? = null
-            return s.split("::").fold(q, {
-                qual, part -> RustQualifiedName(RustNamePart.parse(part), qual)
+            return s.split("::").fold(RustAnonymousId as RustQualifiedName, {
+                qual, part -> RustQualifiedName(RustIdNamePart.parse(part), qual)
             })
         }
 
-        fun toString(qualName: RustQualifiedName): String {
-            val b = StringBuilder()
-
-            val quals = arrayListOf<RustQualifiedName>()
-
-            var qual: RustQualifiedName? = qualName
-            while (qual != null) {
-                quals.add(qual)
-                qual = qual.qualifier
-            }
-
-            quals.asReversed().forEach { q ->
-                val b = StringBuilder()
-
-                b.append(q)
-                b.append("::")
-            }
-
-            return b.toString()
-        }
     }
+
 }
