+++ src/main/kotlin/org/rust/lang/core/names/parts/RustAnonymousIdPart.kt
@@ -1,0 +1,9 @@
+package org.rust.lang.core.names.parts
+
+/**
+ * Name-part of the anonymous 'identifier'
+ */
+object RustAnonymousIdPart : RustNamePart {
+    override val identifier: String
+        get() = "<anonymous>"
+}
